requirements.txt
flask
requests
pymupdf
python-docx
sentence-transformers
torch
numpy